Moved out trace migrations
