+++ src/Bin/OpenEnroth/OpenEnroth.cpp
@@ -23,6 +23,7 @@
 #include "Library/StackTrace/StackTraceOnCrash.h"
 #include "Library/Platform/Application/PlatformApplication.h"
 #include "Library/Trace/EventTrace.h"
+#include "Library/Trace/EventTraceMigrations.h"
 
 #include "Utility/Streams/FileOutputStream.h"
 #include "Utility/String/Format.h"
@@ -70,14 +71,14 @@ void migrateTrace(OpenEnrothOptions::Migration migration, EventTrace *trace) {
     case OpenEnrothOptions::MIGRATION_NONE:
         return;
     case OpenEnrothOptions::MIGRATION_DROP_REDUNDANT_KEY_EVENTS:
-        return EventTrace::migrateDropRedundantKeyEvents(trace);
+        return trace::migrateDropRedundantKeyEvents(trace);
     case OpenEnrothOptions::MIGRATION_COLLAPSE_KEY_EVENTS:
         for (InputAction inputAction : allInputActions())
             if (triggerModeForInputAction(inputAction) != TRIGGER_ONCE)
                 keys.insert(keyboardActionMapping->keyFor(inputAction));
-        return EventTrace::migrateCollapseKeyPressReleaseEvents(keys, trace);
+        return trace::migrateCollapseKeyPressReleaseEvents(keys, trace);
     case OpenEnrothOptions::MIGRATION_DROP_PAINT_AFTER_ACTIVATE:
-        return EventTrace::migrateDropPaintAfterActivate(trace);
+        return trace::migrateDropPaintAfterActivate(trace);
     }
 }
