+++ src/Library/Trace/EventTrace.h
@@ -4,7 +4,6 @@
 #include <string_view>
 #include <memory>
 #include <string>
-#include <unordered_set>
 
 #include "Library/Platform/Interface/PlatformEvents.h"
 #include "Library/Config/ConfigPatch.h"
@@ -59,27 +58,6 @@ struct EventTrace {
 
     static bool isTraceable(const PlatformEvent *event);
     static std::unique_ptr<PlatformEvent> cloneEvent(const PlatformEvent *event);
-
-    /**
-     * Removes all keyboard events that have no effect from a trace. This includes autorepeat events, and key releases
-     * w/o a corresponding key press.
-     *
-     * @param[in, out] trace            Trace to update.
-     */
-    static void migrateDropRedundantKeyEvents(EventTrace *trace);
-
-    /**
-     * Drops key press & release events if they are inside a single frame.
-     *
-     * @param keys                      Set of keys to migrate events for.
-     * @param[in, out] trace            Trace to update.
-     */
-    static void migrateCollapseKeyPressReleaseEvents(const std::unordered_set<PlatformKey> &keys, EventTrace *trace);
-
-    /**
-     * Drops one paint event after each `EVENT_WINDOW_ACTIVATE` and fixes frame timings.
-     */
-    static void migrateDropPaintAfterActivate(EventTrace *trace);
 
     EventTraceHeader header;
     std::vector<std::unique_ptr<PlatformEvent>> events;
